rewrite the implementation and interface in async with lock free
modify the interface of Async.run (add a flag parameter)
modify the parts using `Async.run` in async.c protocol-server.c test-async.c

In this version, the current trouble is the low CPU utilized compared with previous version of async,so
the performance in test case and server-framework remains.
Need some suggesttion and information in advance.